Overhauled config for upcoming modules
Replaced the "both" option with a boolean for each of the 3 modules: bonus random modifier, bonus modifier slot, bonus stats. Added individual config options for each tool stat for the bonus stats module, plus a generic message to play when bonus stats are enabled (empty by default.)

Have not added harvest level and a couple of tool-specific stats as of yet.